Update assignment 04 for GitHub workflow
Majority of work completed to updated assignment 04 on paging
system simulation for GitHub classroom workflows and
evaluation.  Updated and greatly expanded the unit tests in
this assignment.  Modified initial tasks to require students
declare the function signatures and create function documentation.

Added in tests for initial tasks 1-5 to complete the
PagingSystem simulation.  Also added in tests for tasks 6-8 to
complete the ClockPageReplacementScheme class.

Added in hooks for the final system tests as well.

Update all task descriptions in the assignment description documentation.

Update issue templates for the new set of 8+ system test tasks for this
assignment.

Finally update the autograder workflow points for the new set of tasks
and evaluation metrics.